+++ modules/help_browser/src/c/dllMain.c
@@ -20,17 +20,17 @@
 //=============================================================================
 #pragma comment(lib, "shlwapi.lib") // AllocConsole
 #ifdef _DEBUG
-#pragma comment(lib, "boost_system-vc141-mt-gd-1_64.lib")
-#pragma comment(lib, "boost_filesystem-vc141-mt-gd-1_64.lib")
-#pragma comment(lib, "boost_chrono-vc141-mt-gd-1_64.lib")
-#pragma comment(lib, "boost_thread-vc141-mt-gd-1_64.lib")
-#pragma comment(lib, "boost_date_time-vc141-mt-gd-1_64.lib")
+#pragma comment(lib, "boost_system-vc141-mt-gd-1_65_1.lib")
+#pragma comment(lib, "boost_filesystem-vc141-mt-gd-1_65_1.lib")
+#pragma comment(lib, "boost_chrono-vc141-mt-gd-1_65_1.lib")
+#pragma comment(lib, "boost_thread-vc141-mt-gd-1_65_1.lib")
+#pragma comment(lib, "boost_date_time-vc141-mt-gd-1_65_1.lib")
 #else
-#pragma comment(lib, "boost_system-vc141-mt-1_64.lib")
-#pragma comment(lib, "boost_filesystem-vc141-mt-1_64.lib")
-#pragma comment(lib, "boost_chrono-vc141-mt-1_64.lib")
-#pragma comment(lib, "boost_thread-vc141-mt-1_64.lib")
-#pragma comment(lib, "boost_date_time-vc141-mt-1_64.lib")
+#pragma comment(lib, "boost_system-vc141-mt-1_65_1.lib")
+#pragma comment(lib, "boost_filesystem-vc141-mt-1_65_1.lib")
+#pragma comment(lib, "boost_chrono-vc141-mt-1_65_1.lib")
+#pragma comment(lib, "boost_thread-vc141-mt-1_65_1.lib")
+#pragma comment(lib, "boost_date_time-vc141-mt-1_65_1.lib")
 #endif
 //=============================================================================
 int WINAPI DllMain(HINSTANCE hInstance, DWORD reason, PVOID pvReserved)