+++ CHANGELOG.md
@@ -4,6 +4,12 @@ Features:
   [#75](http://github.com/Nelson-numerical-software/nelson/issues/75): Intel Math Kernel Library can used to replace OpenBLAS and FFTW on Windows.
 
 
+Compilation:
+---------
+
+* BOOST 1.65.1 on Windows
+
+
 ## 0.1.10 alpha (2017-10-29)
 
 Features: